requirements.txt
fpdf
openpyxl
openai
streamlit>=1.0
python-dotenv
